requirements.txt
python-telegram-bot==13.7
environs==11.0.0
redis==5.0.0

